requirements.txt
opencv-python-headless
pyzmq
torch
numpy
ultralytics
seaborn
